Split asset manager into multiple files
