+++ internal/assets/manager.go
@@ -18,7 +18,6 @@ package assets
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/hyperledger/firefly/internal/broadcast"
 	"github.com/hyperledger/firefly/internal/config"
@@ -39,8 +38,8 @@ type Manager interface {
 	CreateTokenPool(ctx context.Context, ns, typeName string, pool *fftypes.TokenPool, waitConfirm bool) (*fftypes.TokenPool, error)
 	GetTokenPools(ctx context.Context, ns, typeName string, filter database.AndFilter) ([]*fftypes.TokenPool, *database.FilterResult, error)
 	GetTokenPool(ctx context.Context, ns, typeName, poolName string) (*fftypes.TokenPool, error)
-	GetTokenAccounts(ctx context.Context, ns, typeName, poolName string, filter database.AndFilter) ([]*fftypes.TokenAccount, *database.FilterResult, error)
 	ValidateTokenPoolTx(ctx context.Context, pool *fftypes.TokenPool, protocolTxID string) error
+	GetTokenAccounts(ctx context.Context, ns, typeName, poolName string, filter database.AndFilter) ([]*fftypes.TokenAccount, *database.FilterResult, error)
 	GetTokenTransfers(ctx context.Context, ns, typeName, poolName string, filter database.AndFilter) ([]*fftypes.TokenTransfer, *database.FilterResult, error)
 	NewTransfer(ns, typeName, poolName string, transfer *fftypes.TokenTransferInput) sysmessaging.MessageSender
 	MintTokens(ctx context.Context, ns, typeName, poolName string, transfer *fftypes.TokenTransferInput, waitConfirm bool) (*fftypes.TokenTransfer, error)
@@ -99,138 +98,8 @@ func (am *assetManager) selectTokenPlugin(ctx context.Context, name string) (tok
 	return nil, i18n.NewError(ctx, i18n.MsgUnknownTokensPlugin, name)
 }
 
-func addTokenPoolCreateInputs(op *fftypes.Operation, pool *fftypes.TokenPool) {
-	op.Input = fftypes.JSONObject{
-		"id":        pool.ID.String(),
-		"namespace": pool.Namespace,
-		"name":      pool.Name,
-		"config":    pool.Config,
-	}
-}
-
-func retrieveTokenPoolCreateInputs(ctx context.Context, op *fftypes.Operation, pool *fftypes.TokenPool) (err error) {
-	input := &op.Input
-	pool.ID, err = fftypes.ParseUUID(ctx, input.GetString("id"))
-	if err != nil {
-		return err
-	}
-	pool.Namespace = input.GetString("namespace")
-	pool.Name = input.GetString("name")
-	if pool.Namespace == "" || pool.Name == "" {
-		return fmt.Errorf("namespace or name missing from inputs")
-	}
-	pool.Config = input.GetObject("config")
-	return nil
-}
-
-// Note: the counterpart to below (retrieveTokenTransferInputs) lives in the events package
-func addTokenTransferInputs(op *fftypes.Operation, transfer *fftypes.TokenTransfer) {
-	op.Input = fftypes.JSONObject{
-		"id": transfer.LocalID.String(),
-	}
-}
-
-func (am *assetManager) CreateTokenPool(ctx context.Context, ns string, typeName string, pool *fftypes.TokenPool, waitConfirm bool) (*fftypes.TokenPool, error) {
-	return am.createTokenPoolWithID(ctx, fftypes.NewUUID(), ns, typeName, pool, waitConfirm)
-}
-
-func (am *assetManager) createTokenPoolWithID(ctx context.Context, id *fftypes.UUID, ns string, typeName string, pool *fftypes.TokenPool, waitConfirm bool) (*fftypes.TokenPool, error) {
-	if err := am.data.VerifyNamespaceExists(ctx, ns); err != nil {
-		return nil, err
-	}
-
-	if pool.Key == "" {
-		org, err := am.identity.GetLocalOrganization(ctx)
-		if err != nil {
-			return nil, err
-		}
-		pool.Key = org.Identity
-	}
-
-	plugin, err := am.selectTokenPlugin(ctx, typeName)
-	if err != nil {
-		return nil, err
-	}
-
-	if waitConfirm {
-		requestID := fftypes.NewUUID()
-		return am.syncasync.SendConfirmTokenPool(ctx, ns, requestID, func(ctx context.Context) error {
-			_, err := am.createTokenPoolWithID(ctx, requestID, ns, typeName, pool, false)
-			return err
-		})
-	}
-
-	tx := &fftypes.Transaction{
-		ID: fftypes.NewUUID(),
-		Subject: fftypes.TransactionSubject{
-			Namespace: ns,
-			Type:      fftypes.TransactionTypeTokenPool,
-			Signer:    pool.Key,
-			Reference: id,
-		},
-		Created: fftypes.Now(),
-		Status:  fftypes.OpStatusPending,
-	}
-	tx.Hash = tx.Subject.Hash()
-	err = am.database.UpsertTransaction(ctx, tx, false /* should be new, or idempotent replay */)
-	if err != nil {
-		return nil, err
-	}
-
-	pool.ID = id
-	pool.Namespace = ns
-	pool.TX.ID = tx.ID
-	pool.TX.Type = tx.Subject.Type
-
-	op := fftypes.NewTXOperation(
-		plugin,
-		ns,
-		tx.ID,
-		"",
-		fftypes.OpTypeTokenCreatePool,
-		fftypes.OpStatusPending,
-		"")
-	addTokenPoolCreateInputs(op, pool)
-	err = am.database.UpsertOperation(ctx, op, false)
-	if err != nil {
-		return nil, err
-	}
-
-	return pool, plugin.CreateTokenPool(ctx, op.ID, pool)
-}
-
 func (am *assetManager) scopeNS(ns string, filter database.AndFilter) database.AndFilter {
 	return filter.Condition(filter.Builder().Eq("namespace", ns))
-}
-
-func (am *assetManager) GetTokenPools(ctx context.Context, ns string, typeName string, filter database.AndFilter) ([]*fftypes.TokenPool, *database.FilterResult, error) {
-	if _, err := am.selectTokenPlugin(ctx, typeName); err != nil {
-		return nil, nil, err
-	}
-	if err := fftypes.ValidateFFNameField(ctx, ns, "namespace"); err != nil {
-		return nil, nil, err
-	}
-	return am.database.GetTokenPools(ctx, am.scopeNS(ns, filter))
-}
-
-func (am *assetManager) GetTokenPool(ctx context.Context, ns, typeName, poolName string) (*fftypes.TokenPool, error) {
-	if _, err := am.selectTokenPlugin(ctx, typeName); err != nil {
-		return nil, err
-	}
-	if err := fftypes.ValidateFFNameField(ctx, ns, "namespace"); err != nil {
-		return nil, err
-	}
-	if err := fftypes.ValidateFFNameField(ctx, poolName, "name"); err != nil {
-		return nil, err
-	}
-	pool, err := am.database.GetTokenPool(ctx, ns, poolName)
-	if err != nil {
-		return nil, err
-	}
-	if pool == nil {
-		return nil, i18n.NewError(ctx, i18n.Msg404NotFound)
-	}
-	return pool, nil
 }
 
 func (am *assetManager) GetTokenAccounts(ctx context.Context, ns, typeName, poolName string, filter database.AndFilter) ([]*fftypes.TokenAccount, *database.FilterResult, error) {
@@ -239,247 +108,6 @@ func (am *assetManager) GetTokenAccounts(ctx context.Context, ns, typeName, pool
 		return nil, nil, err
 	}
 	return am.database.GetTokenAccounts(ctx, filter.Condition(filter.Builder().Eq("poolprotocolid", pool.ProtocolID)))
-}
-
-func (am *assetManager) ValidateTokenPoolTx(ctx context.Context, pool *fftypes.TokenPool, protocolTxID string) error {
-	// TODO: validate that the given token pool was created with the given protocolTxId
-	return nil
-}
-
-func (am *assetManager) GetTokenTransfers(ctx context.Context, ns, typeName, name string, filter database.AndFilter) ([]*fftypes.TokenTransfer, *database.FilterResult, error) {
-	pool, err := am.GetTokenPool(ctx, ns, typeName, name)
-	if err != nil {
-		return nil, nil, err
-	}
-	return am.database.GetTokenTransfers(ctx, filter.Condition(filter.Builder().Eq("poolprotocolid", pool.ProtocolID)))
-}
-
-func (am *assetManager) NewTransfer(ns, typeName, poolName string, transfer *fftypes.TokenTransferInput) sysmessaging.MessageSender {
-	sender := &transferSender{
-		mgr:       am,
-		namespace: ns,
-		typeName:  typeName,
-		poolName:  poolName,
-		transfer:  transfer,
-	}
-	sender.setDefaults()
-	return sender
-}
-
-type transferSender struct {
-	mgr          *assetManager
-	namespace    string
-	typeName     string
-	poolName     string
-	transfer     *fftypes.TokenTransferInput
-	sendCallback sysmessaging.BeforeSendCallback
-}
-
-func (s *transferSender) Send(ctx context.Context) error {
-	return s.resolveAndSend(ctx, false)
-}
-
-func (s *transferSender) SendAndWait(ctx context.Context) error {
-	return s.resolveAndSend(ctx, true)
-}
-
-func (s *transferSender) BeforeSend(cb sysmessaging.BeforeSendCallback) sysmessaging.MessageSender {
-	s.sendCallback = cb
-	return s
-}
-
-func (s *transferSender) setDefaults() {
-	s.transfer.LocalID = fftypes.NewUUID()
-}
-
-func (am *assetManager) MintTokens(ctx context.Context, ns, typeName, poolName string, transfer *fftypes.TokenTransferInput, waitConfirm bool) (out *fftypes.TokenTransfer, err error) {
-	transfer.Type = fftypes.TokenTransferTypeMint
-	if transfer.Key == "" {
-		org, err := am.identity.GetLocalOrganization(ctx)
-		if err != nil {
-			return nil, err
-		}
-		transfer.Key = org.Identity
-	}
-	transfer.From = ""
-	if transfer.To == "" {
-		transfer.To = transfer.Key
-	}
-
-	sender := am.NewTransfer(ns, typeName, poolName, transfer)
-	if waitConfirm {
-		err = sender.SendAndWait(ctx)
-	} else {
-		err = sender.Send(ctx)
-	}
-	return &transfer.TokenTransfer, err
-}
-
-func (am *assetManager) BurnTokens(ctx context.Context, ns, typeName, poolName string, transfer *fftypes.TokenTransferInput, waitConfirm bool) (out *fftypes.TokenTransfer, err error) {
-	transfer.Type = fftypes.TokenTransferTypeBurn
-	if transfer.Key == "" {
-		org, err := am.identity.GetLocalOrganization(ctx)
-		if err != nil {
-			return nil, err
-		}
-		transfer.Key = org.Identity
-	}
-	if transfer.From == "" {
-		transfer.From = transfer.Key
-	}
-	transfer.To = ""
-
-	sender := am.NewTransfer(ns, typeName, poolName, transfer)
-	if waitConfirm {
-		err = sender.SendAndWait(ctx)
-	} else {
-		err = sender.Send(ctx)
-	}
-	return &transfer.TokenTransfer, err
-}
-
-func (am *assetManager) TransferTokens(ctx context.Context, ns, typeName, poolName string, transfer *fftypes.TokenTransferInput, waitConfirm bool) (out *fftypes.TokenTransfer, err error) {
-	transfer.Type = fftypes.TokenTransferTypeTransfer
-	if transfer.Key == "" {
-		org, err := am.identity.GetLocalOrganization(ctx)
-		if err != nil {
-			return nil, err
-		}
-		transfer.Key = org.Identity
-	}
-	if transfer.From == "" {
-		transfer.From = transfer.Key
-	}
-	if transfer.To == "" {
-		transfer.To = transfer.Key
-	}
-	if transfer.From == transfer.To {
-		return nil, i18n.NewError(ctx, i18n.MsgCannotTransferToSelf)
-	}
-
-	sender := am.NewTransfer(ns, typeName, poolName, transfer)
-	if waitConfirm {
-		err = sender.SendAndWait(ctx)
-	} else {
-		err = sender.Send(ctx)
-	}
-	return &transfer.TokenTransfer, err
-}
-
-func (s *transferSender) resolveAndSend(ctx context.Context, waitConfirm bool) (err error) {
-	plugin, err := s.mgr.selectTokenPlugin(ctx, s.typeName)
-	if err != nil {
-		return err
-	}
-	pool, err := s.mgr.GetTokenPool(ctx, s.namespace, s.typeName, s.poolName)
-	if err != nil {
-		return err
-	}
-	s.transfer.PoolProtocolID = pool.ProtocolID
-
-	var messageSender sysmessaging.MessageSender
-	if s.transfer.Message != nil {
-		if messageSender, err = s.buildTransferMessage(ctx, s.namespace, s.transfer.Message); err != nil {
-			return err
-		}
-	}
-
-	switch {
-	case waitConfirm && messageSender != nil:
-		// prepare the message, send the transfer async, then send the message and wait
-		return messageSender.
-			BeforeSend(func(ctx context.Context) error {
-				s.transfer.MessageHash = s.transfer.Message.Hash
-				s.transfer.Message = nil
-				return s.Send(ctx)
-			}).
-			SendAndWait(ctx)
-	case waitConfirm:
-		// no message - just send the transfer and wait
-		return s.sendSync(ctx)
-	case messageSender != nil:
-		// send the message async and then move on to the transfer
-		if err := messageSender.Send(ctx); err != nil {
-			return err
-		}
-		s.transfer.MessageHash = s.transfer.Message.Hash
-	}
-
-	tx := &fftypes.Transaction{
-		ID: fftypes.NewUUID(),
-		Subject: fftypes.TransactionSubject{
-			Namespace: s.namespace,
-			Type:      fftypes.TransactionTypeTokenTransfer,
-			Signer:    s.transfer.Key,
-			Reference: s.transfer.LocalID,
-		},
-		Created: fftypes.Now(),
-		Status:  fftypes.OpStatusPending,
-	}
-	tx.Hash = tx.Subject.Hash()
-	err = s.mgr.database.UpsertTransaction(ctx, tx, false /* should be new, or idempotent replay */)
-	if err != nil {
-		return err
-	}
-
-	s.transfer.TX.ID = tx.ID
-	s.transfer.TX.Type = tx.Subject.Type
-
-	op := fftypes.NewTXOperation(
-		plugin,
-		s.namespace,
-		tx.ID,
-		"",
-		fftypes.OpTypeTokenTransfer,
-		fftypes.OpStatusPending,
-		"")
-	addTokenTransferInputs(op, &s.transfer.TokenTransfer)
-	if err := s.mgr.database.UpsertOperation(ctx, op, false); err != nil {
-		return err
-	}
-
-	if s.sendCallback != nil {
-		if err := s.sendCallback(ctx); err != nil {
-			return err
-		}
-	}
-
-	switch s.transfer.Type {
-	case fftypes.TokenTransferTypeMint:
-		return plugin.MintTokens(ctx, op.ID, &s.transfer.TokenTransfer)
-	case fftypes.TokenTransferTypeTransfer:
-		return plugin.TransferTokens(ctx, op.ID, &s.transfer.TokenTransfer)
-	case fftypes.TokenTransferTypeBurn:
-		return plugin.BurnTokens(ctx, op.ID, &s.transfer.TokenTransfer)
-	default:
-		panic(fmt.Sprintf("unknown transfer type: %v", s.transfer.Type))
-	}
-}
-
-func (s *transferSender) sendSync(ctx context.Context) error {
-	out, err := s.mgr.syncasync.SendConfirmTokenTransfer(ctx, s.namespace, s.transfer.LocalID, s.Send)
-	if out != nil {
-		s.transfer.TokenTransfer = *out
-	}
-	return err
-}
-
-func (s *transferSender) buildTransferMessage(ctx context.Context, ns string, in *fftypes.MessageInOut) (sysmessaging.MessageSender, error) {
-	allowedTypes := []fftypes.FFEnum{
-		fftypes.MessageTypeTransferBroadcast,
-		fftypes.MessageTypeTransferPrivate,
-	}
-	if in.Header.Type == "" {
-		in.Header.Type = fftypes.MessageTypeTransferBroadcast
-	}
-	switch in.Header.Type {
-	case fftypes.MessageTypeTransferBroadcast:
-		return s.mgr.broadcast.NewBroadcast(ns, in), nil
-	case fftypes.MessageTypeTransferPrivate:
-		return s.mgr.messaging.NewMessage(ns, in), nil
-	default:
-		return nil, i18n.NewError(ctx, i18n.MsgInvalidMessageType, allowedTypes)
-	}
 }
 
 func (am *assetManager) Start() error {
